Nationial_product/requirements.txt
pyhwpx==0.44.1
langchain==0.3.19
streamlit==1.32.0
python-dotenv==1.0.0
pandas>=2.0.0
requests>=2.0.0
pywin32==308; platform_system=="Windows"
google-generativeai==0.3.1
google-ai-generativelanguage==0.4.0
langchain-core>=0.3.35
langchain-text-splitters>=0.3.6
reportlab==4.3.1
numpy>=1.26.0
pillow>=10.0.0
tqdm>=4.60.0
matplotlib>=3.9.0
scikit-learn>=1.5.0
diskcache>=5.6.0
